+++ src/market.ts
@@ -17,6 +17,8 @@ import * as types from "./types";
 
 import { EventType, OrderbookEvent } from "./events";
 import { Asset } from "./assets";
+import { Product, Strike } from "./program-types";
+import { SubExchange } from "./subexchange";
 
 export class ZetaGroupMarkets {
   /**
@@ -50,6 +52,11 @@ export class ZetaGroupMarkets {
   }
   private _markets: Array<Market>;
 
+  public get perpMarket(): Market {
+    return this._perpMarket;
+  }
+  private _perpMarket: Market;
+
   public set pollInterval(interval: number) {
     if (interval < 0) {
       throw Error("Invalid poll interval");
@@ -64,6 +71,8 @@ export class ZetaGroupMarkets {
   private _lastPollTimestamp: number;
 
   private _subscribedMarketIndexes: Set<number>;
+
+  private _subscribedPerp: boolean;
   /**
    * Returns the market's index.
    */
@@ -146,7 +155,10 @@ export class ZetaGroupMarkets {
   }
 
   public subscribeMarket(marketIndex: number) {
-    if (marketIndex >= this._markets.length) {
+    if (
+      marketIndex >= this._markets.length &&
+      marketIndex != constants.PERP_INDEX
+    ) {
       throw Error(`Market index ${marketIndex} doesn't exist.`);
     }
     this._subscribedMarketIndexes.add(marketIndex);
@@ -154,6 +166,14 @@ export class ZetaGroupMarkets {
 
   public unsubscribeMarket(marketIndex: number): boolean {
     return this._subscribedMarketIndexes.delete(marketIndex);
+  }
+
+  public subscribePerp() {
+    this._subscribedPerp = true;
+  }
+
+  public unsubscribePerp() {
+    this._subscribedPerp = false;
   }
 
   public async handlePolling(
@@ -178,6 +198,18 @@ export class ZetaGroupMarkets {
           }
         })
       );
+
+      if (this._subscribedPerp) {
+        try {
+          await this._perpMarket.updateOrderbook();
+        } catch (e) {
+          console.error(`Orderbook poll failed: ${e}`);
+        }
+        if (callback !== undefined) {
+          let data: OrderbookEvent = { marketIndex: constants.PERP_INDEX };
+          callback(this.asset, EventType.ORDERBOOK, data);
+        }
+      }
     }
   }
 
@@ -198,7 +230,7 @@ export class ZetaGroupMarkets {
         subExchange.zetaGroup.expirySeries.length
     );
 
-    let indexes = [...Array(constants.ACTIVE_MARKETS).keys()];
+    let indexes = [...Array(constants.ACTIVE_MARKETS - 1).keys()];
     for (var i = 0; i < indexes.length; i += constants.MARKET_LOAD_LIMIT) {
       let slice = indexes.slice(i, i + constants.MARKET_LOAD_LIMIT);
       await Promise.all(
@@ -238,6 +270,34 @@ export class ZetaGroupMarkets {
       await sleep(throttleMs);
     }
 
+    // Perps product/market is separate
+    let marketAddr = subExchange.zetaGroup.perp.market;
+    let serumMarket = await SerumMarket.load(
+      Exchange.connection,
+      marketAddr,
+      { commitment: opts.commitment, skipPreflight: opts.skipPreflight },
+      constants.DEX_PID[Exchange.network]
+    );
+    let [baseVaultAddr, _baseVaultNonce] = await getZetaVault(
+      Exchange.programId,
+      serumMarket.baseMintAddress
+    );
+    let [quoteVaultAddr, _quoteVaultNonce] = await getZetaVault(
+      Exchange.programId,
+      serumMarket.quoteMintAddress
+    );
+    instance._perpMarket = new Market(
+      asset,
+      constants.PERP_INDEX, // not in use but technically sits at the end of the list of Products in the ZetaGroup
+      null,
+      types.toProductKind(subExchange.zetaGroup.perp.kind),
+      marketAddr,
+      subExchange.zetaGroupAddress,
+      quoteVaultAddr,
+      baseVaultAddr,
+      serumMarket
+    );
+
     instance.updateExpirySeries();
     return instance;
   }
@@ -271,7 +331,9 @@ export class ZetaGroupMarkets {
    */
   public getMarket(market: PublicKey): Market {
     let index = this.getMarketIndex(market);
-    return this._markets[index];
+    return index == constants.PERP_INDEX
+      ? this._perpMarket
+      : this._markets[index];
   }
 
   /**
@@ -293,6 +355,9 @@ export class ZetaGroupMarkets {
       } else {
         return k;
       }
+    }
+    if (compare(market, this._perpMarket.address) == 0) {
+      return constants.PERP_INDEX;
     }
     throw Error("Market doesn't exist!");
   }
@@ -482,9 +547,13 @@ export class Market {
   }
 
   public updateStrike() {
-    let strike = Exchange.getSubExchange(this.asset).zetaGroup.products[
-      this._marketIndex
-    ].strike;
+    let strike =
+      this._marketIndex == constants.PERP_INDEX
+        ? Exchange.getSubExchange(this.asset).zetaGroup.perp.strike
+        : Exchange.getSubExchange(this.asset).zetaGroup.products[
+            this._marketIndex
+          ].strike;
+
     if (!strike.isSet) {
       this._strike = null;
     } else {